+++ lua/codecompanion/interactions/chat/super_diff.lua
